requirements.txt
# Web框架
Flask==3.1.0
flask-cors==4.0.0

# 数据库
pymongo==4.11.1
dnspython==2.7.0

# 环境变量管理
python-dotenv==1.0.1

# 工具库
blinker==1.9.0
click==8.1.8
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
Werkzeug==3.1.3

# 开发和构建工具
setuptools==75.8.0
wheel==0.45.1
